bugfix: crash when compiling tinyexpr because invariant is broken

+++ src/semantic/const_eval.rs
@@ -177,7 +177,7 @@ impl<'a> ConstEvalCtx<'a> {
             NodeKind::CompoundLiteral(_, init_list) => {
                 if let NodeKind::InitializerList(list) = self.ast.get_kind(*init_list) {
                     let index = if is_real { 0 } else { 1 };
-                    if (index as u16) < list.init_len {
+                    if index < list.init_len {
                         let item = list.init_start.add_offset(index);
                         if let NodeKind::InitializerItem(item) = self.ast.get_kind(item) {
                             return self.eval_float(item.initializer);
